Adapt serialisation scheme such that it is simpler and more comprehensive
i.e. I want to be able to label hosts as not needing passwords to log
in, despite needing a password to sudo.  The current format does not
make it easy to pass that information to a client program in a way
which is easily parsed.  And the current format is a bit too
restricted, complicated and irregular, anyway.